CMP Web A11Y refactoring
